scripts/scrapers/ebookstore_b.py
"""
電子書籍ストアB用のスクレイパー
"""
import logging
import random
import requests
from bs4 import BeautifulSoup
import time
from scripts.scrapers.base import BaseStoreScraper

logger = logging.getLogger(__name__)

class EbookStoreBScraper(BaseStoreScraper):
    """
    電子書籍ストアB用のスクレイパー
    
    実際のスクレイピングロジックはこのクラスに実装します
    """
    
    def _scrape(self):
        """
        電子書籍ストアBからデータをスクレイピングします
        
        Returns:
            list: マンガデータのリスト
        """
        logger.info(f"{self.store.name}からデータのスクレイピングを開始します...")
        
        # 実際には以下のURLにリクエストを送信してデータを取得します
        url = self.store.url
        
        try:
            # ここに実際のスクレイピングコードを実装
            # 現時点ではサンプルデータを返します
            
            # 待機してウェブサーバーに負荷をかけないようにする
            time.sleep(1)
            
            # サンプルデータを生成
            manga_list = self._generate_sample_data()
            
            return manga_list
            
        except Exception as e:
            logger.error(f"スクレイピング中にエラーが発生しました: {str(e)}")
            raise
    
    def _generate_sample_data(self):
        """サンプルデータを生成（実際のスクレイピングの代わり）"""
        logger.info("サンプルデータを生成しています...")
        
        # カテゴリのリスト
        categories = ['all', 'shounen', 'shoujo', 'seinen', 'josei']
        
        # サンプルマンガデータを生成
        manga_data = []
        
        # ストアB用のサンプルタイトル（ストアAとは異なるタイトル）
        titles = [
            "銀河英雄伝説", "鋼の錬金術師", "進撃の巨人", 
            "HUNTER×HUNTER", "NARUTO", "BLEACH",
            "ONE PIECE", "ドラゴンボール", "スラムダンク", "ジョジョの奇妙な冒険"
        ]
        
        authors = [
            "尾田栄一郎", "岸本斉史", "久保帯人", "井上雄彦",
            "荒木飛呂彦", "冨樫義博", "諫山創", "鳥山明",
            "田中芳樹", "荒川弘"
        ]
        
        # ランキング順にデータを生成
        for i in range(len(titles)):
            manga_data.append({
                'title': titles[i],
                'author': authors[i],
                'free_chapters': random.randint(1, 5),
                'free_books': random.randint(0, 2),
                'category_id': random.choice(categories),
                'rank': i + 1  # ランキング (1から始まる)
            })
        
        return manga_data