requirements.txt
Django>=3.2.0,<3.3.0
djangorestframework>=3.12.0,<3.13.0
mysqlclient>=2.0.0,<2.1.0
django-environ>=0.8.0,<0.9.0
drf-yasg>=1.20.0,<1.21.0
django-filter>=2.4.0,<2.5.0
requests>=2.25.0,<2.26.0
beautifulsoup4>=4.9.0,<4.10.0
django-extensions>=3.1.0,<3.2.0